Improve AI logic
Most of the improvements here come down to how the ai chooses restorative actions during combat.
A lot of this is in preparation for combat with more than one combatant on either side, but for now the most noticeable change will be that ai no longer try to heal when already at nearly full health.